+++ builder_pkgs/build_web_compilers/lib/src/build_modules/workers.dart
@@ -8,7 +8,7 @@ import 'dart:math' show min;
 
 import 'package:bazel_worker/driver.dart';
 import 'package:build/build.dart';
-import 'package:build_frontend_server/build_frontend_server.dart';
+import '../build_frontend_server/build_frontend_server.dart';
 import 'package:path/path.dart' as p;
 
 import 'common.dart';