+++ build_runner/test/integration_tests/daemon_command_test.dart
@@ -18,8 +18,6 @@ import 'package:test/test.dart';
 
 import '../common/common.dart';
 
-const defaultTimeout = Timeout(Duration(seconds: 120));
-
 void main() async {
   final webTarget = DefaultBuildTarget((b) {
     b.target = 'web';
@@ -38,7 +36,6 @@ void main() async {
         'build',
         'build_config',
         'build_daemon',
-        'build_modules',
         'build_runner',
         'build_web_compilers',
         'build_test',
@@ -219,5 +216,5 @@ void main() {
     // Check the second client was notified too.
     expect((await results2.next).results.single.status, BuildStatus.started);
     expect((await results2.next).results.single.status, BuildStatus.succeeded);
-  }, timeout: defaultTimeout);
+  });
 }