+++ build_runner/test/integration_tests/web_compilers_test.dart
@@ -10,8 +10,6 @@ import 'package:test/test.dart';
 
 import '../common/common.dart';
 
-const defaultTimeout = Timeout(Duration(minutes: 3));
-
 void main() async {
   test('web compilers', () async {
     final pubspecs = await Pubspecs.load();
@@ -25,7 +23,6 @@ void main() async {
         'build',
         'build_config',
         'build_daemon',
-        'build_modules',
         'build_runner',
         'build_web_compilers',
         'build_test',
@@ -157,5 +154,5 @@ void main() {
           '--define=build_web_compilers:dart_source_cleanup=enabled=true',
     );
     expect(tester.read('root_pkg/build/unused.dart'), null);
-  }, timeout: defaultTimeout);
+  });
 }