+++ builder_pkgs/build_web_compilers/lib/src/dev_compiler_builder.dart
@@ -9,10 +9,10 @@ import 'dart:io';
 import 'package:bazel_worker/bazel_worker.dart';
 import 'package:build/build.dart';
 import 'package:build/experiments.dart';
-import 'package:build_modules/build_modules.dart';
 import 'package:path/path.dart' as p;
 
 import '../builders.dart';
+import 'build_modules/build_modules.dart';
 import 'common.dart';
 import 'errors.dart';