+++ builder_pkgs/build_web_compilers/lib/src/web_entrypoint_marker_builder.dart
@@ -5,9 +5,10 @@
 import 'dart:convert';
 
 import 'package:build/build.dart';
-import 'package:build_modules/build_modules.dart';
 import 'package:glob/glob.dart';
 import 'package:path/path.dart' as p;
+
+import 'build_modules/build_modules.dart';
 
 /// A builder that gathers information about a web target's 'main' entrypoint.
 class WebEntrypointMarkerBuilder implements Builder {