+++ builder_pkgs/build_web_compilers/test/build_modules/module_library_test.dart
@@ -3,9 +3,9 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:build/build.dart';
-import 'package:build_modules/src/module_library.dart';
-import 'package:build_modules/src/platform.dart';
 import 'package:build_test/build_test.dart';
+import 'package:build_web_compilers/src/build_modules/module_library.dart';
+import 'package:build_web_compilers/src/build_modules/platform.dart';
 import 'package:test/test.dart';
 
 void main() {